Integrator self-starting

+++ src/Time/TimeSteppers/AdamsBashforthN.hpp
@@ -9,6 +9,7 @@
 #include <algorithm>
 #include <boost/iterator/transform_iterator.hpp>
 #include <cstddef>
+#include <cstdint>
 #include <iterator>
 #include <map>
 #include <ostream>
@@ -209,7 +210,7 @@ class AdamsBashforthN : public TimeStepper::Inherit {
           history,
       const TimeDelta& time_step) const noexcept;
 
-  size_t number_of_substeps() const noexcept override;
+  uint64_t number_of_substeps() const noexcept override;
 
   size_t number_of_past_steps() const noexcept override;
